# Conflicts:
#	src/main/java/ru/yandex/practicum/filmorate/service/film/FilmService.java

+++ src/main/java/ru/yandex/practicum/filmorate/model/ReviewLike.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @AllArgsConstructor
 @Builder
 public class ReviewLike {
-    private Integer reviewId;
-    private Integer userId;
+    private Long reviewId;
+    private Long userId;
     private Boolean like;
 }
